# book-network-backend/src/main/java/com/example/demo/authentication/AuthenticationResponse.java
package com.example.demo.authentication;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}