# book-network-backend/src/main/java/com/example/demo/book/BookController.java
package com.example.demo.book;

import com.example.demo.common.PageResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;

@RestController
@RequestMapping("books")
@Tag(name="Book")
@RequiredArgsConstructor
public class BookController {
    private final BookService bookService;
    @PostMapping()
    public ResponseEntity<Integer> saveBook(@Valid @RequestBody BookRequest request, Authentication currentUser){
        return ResponseEntity.ok(bookService.save(request,currentUser));
    }
    @GetMapping("/{book-id}")
    public ResponseEntity<BookResponse> getBook(@PathVariable("book-id") Integer id){
        return ResponseEntity.ok(bookService.findById(id));

    }
    @GetMapping
    public ResponseEntity<PageResponse<BookResponse>> findAllBooks(
            @RequestParam(name= "page",defaultValue = "0",required = false) int page,
            @RequestParam(name = "size",defaultValue = "10",required = false) int size,
            Authentication connectedUser
    ){
        return ResponseEntity.ok(bookService.findAllBooks(page,size,connectedUser));
    }
    @GetMapping("/owner")
    public ResponseEntity<PageResponse<BookResponse>>findAllBooksByOwner(
            @RequestParam(name= "page",defaultValue = "0",required = false) int page,
            @RequestParam(name = "size",defaultValue = "10",required = false) int size,
            Authentication connectedUser
    ){
        return ResponseEntity.ok(bookService.findAllBooksByOwner(page,size,connectedUser));
    }
    @GetMapping("/borrowed")
    public ResponseEntity<PageResponse<BorrowedBookResponse>>findAllBorrowedBooks(
            @RequestParam(name= "page",defaultValue = "0",required = false) int page,
            @RequestParam(name = "size",defaultValue = "10",required = false) int size,
            Authentication connectedUser
    ){
        return ResponseEntity.ok(bookService.findAllBorrowedBooks(page,size,connectedUser));
    }
    @GetMapping("/returned")
    public ResponseEntity<PageResponse<BorrowedBookResponse>> findAllReturnedBooks(
            @RequestParam(name = "page", defaultValue = "0", required = false) int page,
            @RequestParam(name = "size", defaultValue = "10", required = false) int size,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(bookService.findAllReturnedBooks(page, size, connectedUser));
    }
    @PatchMapping("/shareable/{book-id}")
    public ResponseEntity<?> updateShareableStatus(@PathVariable("book-id") Integer bookId,Authentication connectedUser){
        return ResponseEntity.ok(bookService.updateShareableStatus(bookId,connectedUser));

    }
    @PatchMapping("/archived/{book-id}")
    public ResponseEntity<?> updateArchivedStatus(@PathVariable("book-id") Integer bookId,Authentication connectedUser){
        return ResponseEntity.ok(bookService.updateArchivedStatus(bookId,connectedUser));

    }
    @PostMapping("/borrow/{book-id}")
    public ResponseEntity<?> borrowBook(@PathVariable("book-id") Integer bookId,Authentication connectedUser){
        return ResponseEntity.ok(bookService.borrowBook(bookId,connectedUser));

    }
    @PatchMapping("borrow/return/{book-id}")
    public ResponseEntity<Integer> returnBorrowBook(
            @PathVariable("book-id") Integer bookId,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(bookService.returnBorrowedBook(bookId, connectedUser));
    }
    @PatchMapping("borrow/return/approve/{book-id}")
    public ResponseEntity<Integer> approveReturnBorrowBook(
            @PathVariable("book-id") Integer bookId,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(bookService.approveReturnBorrowedBook(bookId, connectedUser));
    }
    @PostMapping(value = "/cover/{book-id}",consumes = "multipart/form-data")
    public ResponseEntity<?>uploadBookCoverPicture(@PathVariable("book-id") Integer bookId, @RequestPart("file")MultipartFile file, Authentication currentUser){
        bookService.uploadCoverPicture(bookId,file,currentUser);
        return ResponseEntity.ok().build();
    }
    @GetMapping(value = "getAll")
    public ResponseEntity<List<BookResponse>> getAllBooks(){
       List<BookResponse> books =  bookService.getAllBooks();
       return ResponseEntity.ok(books);
    }

}