# book-network-backend/src/main/java/com/example/demo/feeback/FeedBackResponse.java
package com.example.demo.feeback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder

public class FeedBackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}